backend/requirements.txt
# FastAPI y servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# IA y Audio
openai-whisper>=20231117
librosa>=0.10.0
scikit-learn>=1.3.0
xgboost>=2.0.0
google-generativeai>=0.3.0
google-cloud-texttospeech>=2.14.0
pydub>=0.25.1

# Base de datos
motor>=3.3.0
pymongo>=4.6.0
pydantic>=2.5.0

# Utilidades
numpy>=1.24.0
scipy>=1.11.0
